requirements.txt
shiny
seaborn
pandas
pyserial
matplotlib
